Day 6 Solution
Used a u32 as a bit set to store a flag for each letter a-z (26 options). There is a vector to store each group and a vector to store each person in the group.

For part 1 just needs to or each bitset of a group to get the numer of yes(s) per group, then sum the groups for the answer.

For part 2, we just need to and each bitset together then sum the groups for the answer.