Day 10 Solution
Really happy with this day as it was the last day I completed on day of its release and it was my best results.
Day       Time   Rank  Score       Time   Rank  Score
 10   00:11:04   2776      0   00:33:05   1615      0

Solution is very simple,, most of the work is done in the parser. The parser simplifies the problem by sorting all the voltages and adding in the assumed start and end points to our sorted list. All part 1 needs to do is compute all the differences and count the 1 volt and 3  volt differences. Part 2 uses a cache backed by a hashmap (could be a vec) to sum all configs of voltages that are within 3 volts. This is a pretty simple dynamic programming solution, but was super quick to write with all the implemented functions from rust std lib. Could be cleaned up a bit, but overall happy with the speed I was able to write at.