Day 9 Solution
Took a straight forward solution to both parts.

Part 1 just has a sliding window that goes through every pair of values in the window to find a sum equal to the 26th value.

Part 2 is also very brute force, but at least has the minor optimization of iteratively building the sum instead of computing it every increase in window. This could be further optimized by resusing the sum from the previous window. This would require then either shrinking or growing the window depending on if the dropped value brings the sum below the target value.