Day 17 Solution
Tried using nested vectors and expanding the map each cycle, but couldn't resolve the lifetime issues easily.

So reverted to just using hashmap/set with key being the point of active points. Adding all the neighbors of active point into a hashmap make it easy to construct the next steps active points.

Would be nice to refactor this to be more generic over the number of dimensions.